# New state tracker and descriptor set implementation