--- a/MoltenVK/MoltenVK/Commands/MVKCmdDispatch.mm
+++ b/MoltenVK/MoltenVK/Commands/MVKCmdDispatch.mm
@@ -55,7 +55,7 @@ void MVKCmdDispatch::encode(MVKCommandEncoder* cmdEncoder) {
 		} else {
 			// We have to pass the base group in a buffer.
 			uint32_t base[3] = {(uint32_t)mtlThreadgroupCount.origin.x, (uint32_t)mtlThreadgroupCount.origin.y, (uint32_t)mtlThreadgroupCount.origin.z};
-			cmdEncoder->setComputeBytes(mtlEncoder, base, sizeof(base), pipeline->getIndirectParamsIndex().stages[kMVKShaderStageCompute]);
+			cmdEncoder->setComputeBytes(mtlEncoder, base, sizeof(base), pipeline->getImplicitBuffers().ids[MVKImplicitBuffer::DispatchBase]);
 		}
 	}
 	[mtlEncoder dispatchThreadgroups: mtlThreadgroupCount.size