CTS run, no regressions.
<details>
<summary>Newly passing tests</summary>

```sh
dEQP-VK.api.descriptor_pool.out_of_pool_memory
dEQP-VK.multiview{.dynamic_rendering,.renderpass2,}.view_mask_iteration.*
dEQP-VK.pipeline.monolithic.extended_dynamic_state.before_good_static.stencil_state_extra_pipeline_face_*
```
</details>

Sorry for the massive PR.  I could try to split it between the state tracker and descriptor set changes, but due to the design differences between the new state tracker and the old descriptor set implementation, the code sticking them together is very temporary and probably pretty slow.  (Also push descriptors aren't supported at that point, though they could be made to work, once again very slowly.)  I could also move the initial commits that just add helper classes and methods to MVKFoundation to a separate PR if that's preferred.

### Highlights

- Much faster
  - O(1) descriptor pool reset for pools with freeing disabled
  - Descriptor set copy is memcpy, no refcounting
  - Allocating descriptor sets will never allocate MTLBuffers / other memory
- More resilient to desyncing between our view of the Metal command encoder and what's actually there
- Should be much easier to add alternate descriptor layouts if we e.g. wanted to support supplying external Metal shaders using the descriptor layout used by Apple's DXIL shader converter

### Implementation Details

State tracking is split into a Vulkan state, which represents the stuff the application has requested of us, and a Metal state, which is what we've bound to Metal.  Vulkan binding functions only ever modify the Vulkan state and the dirty flags in the Metal state, never the actual data in the Metal state.  `finalizeDrawState` and `finalizeDispatchState` are in charge of binding stuff to Metal to match the current Vulkan state.  This is all done through `bindMetalResources`, which takes a Vulkan stage and a Metal stage to sync up, so the pipelines don't have to worry about e.g. whether they should be binding their vertex or tess eval bindings.  The shared code also means that things like e.g. applying vertex strides to tess-in-compute just works, since it uses the same function to bind as the one used to bind vertex buffers to a normal vertex stage.  Resource dirty tracking is now done per-resource, and API draws and helper draws go through the same resource tracking methods, both removing the need for a special override bit to indicate a non-API resource and the need for API resources to be manually tracked, which made it incredibly easy to mess up pipeline layout compatibility.  This does mean that any helper draws / dispatches must use the state tracker methods to bind their resources to avoid desync, but in return they no longer force the API resources to be fully rebound, only what's necessary will be rebound.  The binding list is now also properly cleared when starting a new encoder, so we don't bind useless resources that were used on the previous encoder at now-unused binding indices to the new one.

Descriptor sets and descriptors are now pure, non-reference-counted data.  To facilitate this, resource liveness handling has been moved to a dictionary on the MVKDevice containing all the live resources, rather than by having descriptors reference count their resources so they stay around and then nulling out the Metal resources inside of those on API destruction.  Unextended vulkan requires that applications ensure any statically used resources are kept alive during execution, so in that case, we don't consult the liveness dictionary at all and just assume resources are kept alive.  (Unlike the old state tracker, the new one only attempts to bind statically used resources.)  In the case an application uses PARTIALLY_BOUND_BIT, we consult the liveness dictionary to make sure resources are still alive before binding them (unless residency sets are in use, in which case we get partially bound for free).  This is done with proper locking, unlike the previous implementation, so there should be no more race conditions using PARTIALLY_BOUND, just worse performance.  Since this makes MoltenVK much more strict with resource liveness, so I've also added an MVK_CONFIG_LIVE_CHECK_ALL_RESOURCES config option that non-compliant applications can use to get the old behavior.

Moving liveness to a side structure means that anyone who doesn't require liveness checks doesn't have to pay for them.  The following use cases get full speed with no liveness checking at all:
- Unextended Vulkan on MVK with argument buffers disabled or Metal 3 argument encoding
- Vulkan with partially bound enabled on MVK with residency sets

Argument encoders need liveness checking on descriptor copy operations, because descriptor copy is implemented by re-encoding the resources, and partially bound needs liveness checking to decide whether to UseResource resources.

Descriptor pools now allocate all GPU descriptor data from the same buffer.  For cases where MoltenVK previously used temporary buffers (e.g. descriptor size buffers and inline uniforms), all buffers are allocated from the same main buffer and internal pointers are added to point to the other sections of data.  This means that we don't have to track extra buffers and we also don't have to separately useResource the descriptor size buffer, since it's the same buffer as the main one.  I've also removed the ability to allocate new descriptors even after running out of space.  This previously caused massive performance drops if you hit it (e.g. dropping to 2fps in RDR2 on vkd3d), and could also break applications that relied on getting an OOM error to reset pools.  The new implementation isn't fussy about what types of resource you allocate (e.g. if you create a pool with space for 32 sampled images and 0 buffers and later try to allocate a buffer out of it, it will succeed), which brings it in line with most other Vulkan implementations I know of, so I doubt the lack of ability to have pools grow infinitely will be a big issue.

Since descriptor sets are now just buffers of raw data, their layout is now handled entirely by descriptor set layouts.  The descriptor set layout constructor makes all the decisions about how to lay out the contents of the descriptor set, which it encodes into an array of MVKDescriptorBindings which contain things like the offset in the cpu buffer, offset in the gpu buffer, and what format the descriptor is stored as.  From there, everything else consults those bindings to interpret the data in the descriptor set, which means that if you wanted to create an MSL shader extension that let the application request specific argument buffer layouts for its descriptor sets, all you would need to do is convert that request to MVKDescriptorBindings and the rest would be handled automatically, even if the requested layout looked nothing like what MoltenVK would use for SPIRV-Cross-generated shaders.  (Note: If you do want to create such an extension, please don't expose any of MVK's SoA descriptor formats, as I don't think we will want to keep support for those around forever, especially since they're completely incompatible with what we would need to implement VK_EXT_descriptor_buffer.)

Push descriptor sets are now just descriptors sets whose backing buffer is managed by the command encoder.  We no longer need two copies of every bind method, one to bind directly to Metal and one to write the descriptor into a descriptor set.  And no need to for push descriptors worry about what Metal stages to bind to either, they just place their resources in the Vulkan state and let `bindMetalResources` figure out where those resources actually need to go.  Same for push constants.

### Benchmarks

I tested RDR2 on vkd3d, a NFS Carbon gsdump on an old pre-push-descriptor build of PCSX2 (v1.7.4886), a newer build of PCSX2 that uses push descriptors (v2.2.0), and a Metroid Prime 2 map fifolog in Dolphin, all of which are very heavy on the CPU.  Testing was done on a 2019 iMac with an i9-9900K and Radeon Pro 580X, comparing the new implementation to the MoltenVK it was originally written against, from a few months ago.  RDR2 encodes commands on 8 threads, and submits on a separate thread which ends up being the main bottleneck.  Each draw binds a fairly large number of resources.  PCSX2 and Dolphin both encode on one thread and were configured to submit on the same thread, so I could compare both MoltenVK's asynchronous and synchronous submission modes.  PCSX2 uses 1-2 resources per draw, while Dolphin uses 18.  (Async for both is bottlenecked on the Vulkan encode thread, while sync is bottlenecked on encode + submit since they both happen on the same thread.)  RDR2 was additionally tested with both LIVE_CHECK_ALL_RESOURCES enabled and disabled to get an idea of how much overhead the liveness checking has.  Dolphin was not tested with argument buffers because the old MVK build I was using was from when those were broken in Dolphin.

|          Test         |Before|After|% Diff|
|:----------------------|-----:|----:|-----:|
| RDR2 Bind             |   28 |  38 | 136% |
| RDR2 Bind LiveCheck   |  N/A |  35 | 125% |
| RDR2 Metal3 ArgBuf    |   29 |  35 | 120% |
| RDR2 Arg Encoder      |   22 |  35 | 159% |
| RDR2 ArgEnc LiveCheck |  N/A |  32 | 110% |
| PCSX2 Bind Async      |  108 | 122 | 113% |
| PCSX2 Metal3 Async    |  103 | 118 | 115% |
| PCSX2 ArgEnc Async    |   39 | 115 | 295% |
| PCSX2 Push Async      |  109 | 109 | 100% |
| PCSX2 Bind Sync       |   78 |  98 | 126% |
| PCSX2 Metal3 Sync     |   70 |  93 | 142% |
| PCSX2 ArgEnc Sync     |   33 |  93 | 281% |
| PCSX2 Push Sync       |   78 |  89 | 114% |
| Dolphin Bind Async    |  146 | 160 | 110% |
| Dolphin Bind Sync     |   99 | 120 | 121% |

### Future Work

- Descriptor update templates could use some love.  At least store the direct binding index instead of the Vulkan binding ID (which might require a binary search to find the binding).  I don't know of any applications that use update templates, so I didn't really touch them.
- I was originally considering trying to dirty track which resources were bound that matched the current pipeline layout to avoid having to rescan them all if an application issues multiple draws without binding new descriptor sets, but performance was good enough without that and the more dirty tracking you do the easier it is to desync.  If someone finds an application that spends a lot of time in bindMetalResources without actually calling Metal bind functions (an indication that it's scanning lots of resources that are already bound), we could revisit that.
- I see a lot of time being spent in things like getMTLTexture() during descriptor set updates.  It might make sense to stop trying to lazily allocate MTLTextures and copy them to the view objects during view creation, so we don't need to go through multiple pointers and method calls to get each one.
- We should be able to skip the CPU side buffer for descriptors entirely when residency sets are in use.
- Look into why argument buffers are consistently slower than binding.