@squidbus 2025-07-29
This seems to fix the state tracking issues I have in https://github.com/KhronosGroup/MoltenVK/issues/2586