@V380-Ori 2025-08-20
I am just a layman so my technical knowledge any of those changes are extremely limited or non-existent at all.

This PR causes **immediate crash in Ryujinx emulator** for most of the games I tested.

### Context:

- Ryujinx is no longer in development since October 2024.
- Ryujinx might violate Vulkan spec for emulation but work on NVIDIA/AMD/Intel drivers so this might be not PR issue at all, I don’t want waste someone's time.

Can someone please take a look at this crash trace?

[Ryujinx_MoltenVK-2572.txt](https://github.com/user-attachments/files/21900569/Ryujinx_MoltenVK-2572.txt)

### Additional (possibly off-topic) info:

Ryujinx had to stay on **MoltenVK 1.2.0** (2022) due to several issues. The most crucial one: without #2427, nothing displays on screen since the introduction of Metal 3 argument buffers.

Other changes Ryujinx needed:
* Remove _hasRasterized check in endMetalRenderPass (fixes occlusion query issue #1842).
* Revert: “Add support for VK_EXT_external_memory_host extension” (#1854).
* Revert: “Add VK_EXT_external_memory_metal” (#2314).
* Revert: “Add support for VK_EXT_index_type_uint8” (#2504).
* Revert: “Remove CPU path from Uint8 index handling” (#2506).
* Change maxPushDescriptors property to 16.

Note, there was no changes made in MoltenVK for this PR test.

Thanks to everyone who contributes to this awesome project!