@V380-Ori 2025-08-21
I just want to clarify: without these changes, Ryujinx still runs, but it shows a black screen and some graphical issues. With this PR, however, it immediately crashes.

I’m not expecting this to be fixed specifically for Ryujinx, nor am I trying to delay the merge of this PR — this might very well be an application-side issue.

Sorry for the noise on this topic.