@MrGcGamer 2025-08-21
kinda off-topic: ^true, but I think it would be really cool to have a new build of MoltenVK working with Ryujinx again, as I think that would require a bit of reworking / fixing with metal argument buffers, as I also have another application which is forced to stay on old MVK versions as the argument buffers completely breaks it.

Also for the noise, just wanted to get that out there as well..