@V380-Ori 2025-08-22
@etang-cw Yes, that fixed it! Thank you so much!

About “_hasRasterized” — as far as I know only Unreal Engine games for Switch have the issue without removing it (including this PR):


https://github.com/user-attachments/assets/3c814e69-5f5a-4348-942e-5868c0fb6493



I am not asking it to be removed as Ryujinx might violate Vulkan spec.

Again, thank you!