@V380-Ori 2025-08-30
@etang-cw It turns out that there were still a couple of games remaining that were crashing with the exact same crash trace, I should’ve tested more earlier… sigh.

Since I couldn’t get gfxreconstruct to work, I had to poke around and found that removing [the line that sets “_dirty” to false in “MVKOcclusionQueryCommandEncoderState::encode”](https://github.com/KhronosGroup/MoltenVK/blob/565e87997a75978d792216371e6e0c91d99d86ed/MoltenVK/MoltenVK/Commands/MVKCommandEncoderState.mm#L1905) seems to resolve this!

@MrGcGamer The video above is Ryujinx with MoltenVK based on this PR branch, with the changes I mentioned, but specifically without removing “_hasRasterized” check as author of this PR was curios.

Note, that you cannot simply drag-n-drop libMoltenVK.dylib into Ryujinx, as null descriptors has to be disabled although this PR fixed it.

Another thing, #2427  is no longer can be applied with this PR, so Ryujinx has to be changed to not use batched descriptor sets for writing images.

About performance, I did a quick comparison of the game “Stray” with master and this PR: saved 3ms of GPU time!

<img width="1280" height="800" alt="ryujinx-stray-master" src="https://github.com/user-attachments/assets/f4612732-0378-438f-81af-a6d461f212c0" />

<img width="1280" height="800" alt="ryujinx-stray-pr" src="https://github.com/user-attachments/assets/895e23f4-e8bb-493f-89e2-1508c4b851d8" />