@etang-cw 2025-09-02
> Since I couldn’t get gfxreconstruct to work, I had to poke around and found that removing [the line that sets “_dirty” to false in “MVKOcclusionQueryCommandEncoderState::encode”](https://github.com/KhronosGroup/MoltenVK/blob/565e87997a75978d792216371e6e0c91d99d86ed/MoltenVK/MoltenVK/Commands/MVKCommandEncoderState.mm#L1905) seems to resolve this!

Looking more at occlusion queries, it looks like an implementation for occlusion queries within a single render pass that someone slapped a "works across multiple render passes" label on even though it actually doesn't.  I've opened #2604 to fix that, if you could test Ryujinx with it.

> Another thing, #2427 is no longer can be applied with this PR, so Ryujinx has to be changed to not use batched descriptor sets for writing images.

That was working and it looks like I broke it in a cleanup.  Should be working again.