@V380-Ori 2025-09-02
@etang-cw

Here’s Xcode gpu traces (1.94 GB after decompression): [Google Drive](https://drive.google.com/file/d/1QXFzDXf0jzyPRGKWml-Sln7OeoPnugkW/view?usp=share_link)

I really tried gfxreconstruct but Ryujinx just crashes when launching via Vulkan loader, while tools like gpuinfo from SDK worked. If Xcode gpu trace is not helpful, I will spend more time figuring out why gfxreconstruct crashes with Ryujinx…