@squidbus 2025-09-02
I don’t know why it closed, mobile GitHub is being strange.

But anyway, for gfxreconstruct with the Vulkan loader, are you making sure to replace the system libMoltenVK with your new build, so that it loads the correct one?