@V380-Ori 2025-09-03
> But anyway, for gfxreconstruct with the Vulkan loader, are you making sure to replace the system libMoltenVK with your new build, so that it loads the correct one?

I tried so many things, but no matter what its crashing only when using “VK_LAYER_LUNARG_gfxreconstruct” layer, validation layer "VK_LAYER_KHRONOS_validation" works...

`00:00:00.749 |E| .NET TP Worker [Vulkan Loader] INFO | LAYER:   Inserted device layer "VK_LAYER_LUNARG_gfxreconstruct" (/usr/local/share/vulkan/explicit_layer.d/../../../lib/libVkLayer_gfxreconstruct.dylib)
00:00:00.800 |E| .NET TP Worker Fatal error.
00:00:00.800 |E| .NET TP Worker System.AccessViolationException: Attempted to read or write protected memory. This is often an indication that other memory is corrupt.`