@V380-Ori 2025-09-03
@etang-cw Yes! That fixed it, thank you! I am sorry that I couldn’t provide gfxreconstruct so it would be easier…